+++ pom.xml
@@ -36,7 +36,7 @@
                 <artifactId>javafx-maven-plugin</artifactId>
                 <version>0.0.4</version>
                 <configuration>
-                    <mainClass>com.sobermind.works.NotMain</mainClass>
+                    <mainClass>com.sobermind.works.App</mainClass>
                 </configuration>
                 <executions>
                     <execution>
