+++ src/main/java/com/sobermind/works/Controller.java
@@ -8,6 +8,10 @@ import java.io.*;
 import java.net.URL;
 import java.util.*;
 
+/**
+ * Authors: Davidx3D & Asandax6
+ */
+
 public class Controller implements Initializable{
 
     @FXML private CheckBox offscrn,stay,offcl,showt,fsmode,shortmod,cmdOut;
@@ -16,11 +20,12 @@ public class Controller implements Initializable{
     @FXML private Label iplabel,applicable,portlabel;
     @FXML private RadioButton wired,wireless,bt,snd;
     @FXML private ListView lists;
+    @FXML private Label lbl;
 
     //Button Functions
     @FXML
     void onRunbtn() throws Exception {
-
+        //Variable Declaration Below
         boolean Off = offscrn.isSelected();
         boolean On = stay.isSelected();
         boolean Offcl = offcl.isSelected();
@@ -47,26 +52,28 @@ public class Controller implements Initializable{
         if(Fs) cmd.add(" -f");
         if(Short) cmd.add(" --shortcut-mod=lalt");
         if (lists.getSelectionModel().selectedItemProperty().equals("")) {
-
+        //Line 54 To 58 Is Buggy, Will Fix When I Have Free Time From Other Projects
         } else {
             cmd.add(" -s"+lists.getSelectionModel().getSelectedItem());
         }
         if(Bit.matches("[0-9]+") && Bit.length() < 3) cmd.add("-b "+Bit+"M");
         if(Max.matches("[0-9]+") && Max.length() > 2) cmd.add("--max-size "+Max);
-
+        //Code Below Removes The Symbols That Interfere With The Scripts From The Array
         String Script = cmd.toString()
                               .replace(",", "")
                               .replace("[", "")
                               .replace("]", "")
                               .trim();
 
-        //Below Is The Important Stuff
+        //Below Is The Important Stuff Which Runs The Scrcpy App With The Scripts From The Array
         if (Cmd){
+            //Opens Cmd And Runs Scrcpy With Added Scripts (Without Printing To Output TextArea)
             Process p = Runtime.getRuntime().exec("cmd.exe /c scrcpy " + Script + " "+Adv);
-            //Opens Cmd And Runs Scrcpy With Added Scripts
+            BufferedReader r = new BufferedReader(new InputStreamReader(p.getInputStream()));
         } else {
+            //Opens Cmd And Runs Scrcpy With Added Scripts (While Printing To Output TextArea)
             Process p = Runtime.getRuntime().exec("cmd.exe /c scrcpy " + Script + " "+Adv);
-            //Opens Cmd And Runs Scrcpy With Added Scripts
+            //Below Is The Code That Reads The Output From Cmd After Running Scrcpy And Display It On a TextArea
             String sr = null;
             BufferedReader rs = new BufferedReader(new InputStreamReader(p.getInputStream()));
             while ((sr = rs.readLine()) != null) {
@@ -80,13 +87,14 @@ public class Controller implements Initializable{
         }
     @FXML
     protected  void onExitbtn(){
-
+        //Will Exit App If Exit Button Is Clicked
         System.exit(0);
     }
     public void onWired(ActionEvent actionEvent) {
         if (wired.isSelected()) {
             wireless.setSelected(false);
         }
+        //The Simple Code Below Will Allow The Hidden Wireless Functions To Appear And Be Usable
         applicable.setVisible(false);
         devip.setVisible(false);
         iplabel.setVisible(false);
@@ -114,22 +122,35 @@ public class Controller implements Initializable{
     //The Code Below Uses IDevice To Check If a Device Has Been Connected Or Disconnected From The PC
     @Override
     public void initialize(URL location, ResourceBundle resources) {
-
+        //Line 126 To 128 Currently Not Working
+        if (devices.getText() != ""){
+            lbl.setText("");
+        }
+        //The Below Are Set To False As They Are Options For Wireless And Wired Is Default
+        wired.setSelected(true);
+        applicable.setVisible(false);
+        devip.setVisible(false);
+        iplabel.setVisible(false);
+        port.setVisible(false);
+        portlabel.setVisible(false);
+        //Uhm....Beep...Bop...Boop
         AndroidDebugBridge.init(false);
         devices.setEditable(false);
         output.setEditable(false);
+        cmdOut.setSelected(true);
 
         AndroidDebugBridge debugBridge = AndroidDebugBridge.createBridge("adb.exe", true);
         if (debugBridge == null) {
             System.exit(1);
         }
+        //The Code Below Uses ddmlib To Check For Inserted Adb Devices On The Computer And Display The Serial Number
         AndroidDebugBridge.addDeviceChangeListener(new AndroidDebugBridge.IDeviceChangeListener() {
 
             public void deviceChanged(IDevice device, int arg1) {
                 //Not Necessary For Now
             }
             public void deviceConnected(IDevice device) {
-
+                //When a Device Is Connected The App Will Run The Command Adb Devices To Check Connected Adb Devices
                 try {
                     devices.setText("");
                     String ss = null;
@@ -143,17 +164,15 @@ public class Controller implements Initializable{
                         devices.setEditable(false);
                         devices.setWrapText(true);
                     }
-                    while ((ss = r.readLine()) != null) {
-                        System.out.println(ss);
-                    }
                 }
                 catch(Exception e) {
                     e.printStackTrace();
                 }
+                //Code Below Takes Device Serial Number From Adb And Places Inside a List View
                 lists.getItems().add(device.getSerialNumber());
             }
             public void deviceDisconnected(IDevice device) {
-
+                //When a Device Is Disconnected It Will Be Removed From The List View Of Available Devices
                 lists.getItems().remove(device.getSerialNumber());
                 devices.setText("");
                 devices.setText(String.format("%s Disconnected",device.getSerialNumber()));
@@ -175,26 +194,37 @@ public class Controller implements Initializable{
                 }
             }
         });
-        wired.setSelected(true);
-        applicable.setVisible(false);
-        devip.setVisible(false);
-        iplabel.setVisible(false);
-        port.setVisible(false);
-        portlabel.setVisible(false);
     }
     public void onBT(ActionEvent actionEvent) {
+        //When BT Radio Button Is Selected The Sndcpy Radio Button Will Be Off
+        //And If The Confirmation Is Yes The It Will Run BT Audio Connector
         if (bt.isSelected()){
             snd.setSelected(false);
             Audio.BTCon();
         }
     }
     public void onSndcpy(ActionEvent actionEvent) {
+        //When Sndcpy Radio Button Is Selected The BT Radio Button Will Be Off
+        //And If The Confirmation Is Yes The It Will Run Sndcpy
         if (snd.isSelected()){
             bt.setSelected(false);
             Audio.SndCon();
              }
         }
-    public void onTest(ActionEvent actionEvent) {
 
+    public void onDisp(ActionEvent actionEvent) {
+        //This Is For The Alert When The Disable Display Out CheckBox Is Toggled To Off
+        Alert Disp = new Alert(Alert.AlertType.CONFIRMATION);
+        Disp.setTitle("Disable Output Page");
+        Disp.setHeaderText("The App Will Hang When This Is Disabled");
+        Disp.setContentText("The App Will Run Scrcpy Then Hang (It Will Still Work, Just Not Well)");
+
+        Optional<ButtonType> result = Disp.showAndWait();
+        if (result.get() == ButtonType.OK){
+           cmdOut.setSelected(false);
+        }
+        else {
+            cmdOut.setSelected(true);
+        }
     }
 }
