+++ src/main/java/com/sobermind/works/Main.java
@@ -1,0 +1,13 @@
+package com.sobermind.works;
+
+/**
+ * Authors: Davidx3D & Asandax6
+ */
+
+public class Main {
+
+    public static void main (String [] args){
+        //This Class Is Just Here To Make Sure The App Builds Since I Get Errors When The Main Class Is Extended
+        App.main(args);
+    }
+}
