--- src/main/java/com/sobermind/works/PsuedoMain.java
@@ -1,9 +1,0 @@
-package com.sobermind.works;
-
-public class PsuedoMain {
-
-    public static void main (String [] args){
-        //This Class Is Just Here To Make Sure The App Builds Since I Get Errors When The Main Class Is Extended
-        NotMain.main(args);
-    }
-}
